backend/requirements.txt
# === BACKEND CORE ===
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiohttp==3.12.13
python-dotenv==1.0.1
beautifulsoup4==4.12.3

# === OCR + NLP ===
pytesseract==0.3.10
Pillow==10.2.0
textblob==0.19.0
nltk==3.9.1

# === SUMMARIZATION ===
transformers==4.54.0
torch==2.2.0
